refactor the model and route
