+++ app/Http/Controllers/PostController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Post;
-use App\PostImage;
+use App\Models\Post;
+use App\Models\PostImage;
 use Auth;
 use Storage;
+use Illuminate\Support\Facades\DB;
 
 class PostController extends Controller
 {
